Autonomost ubuntu fix (#1)
* Ubuntu 20.04 install issue fix

when cloning this repo the installer script isn't executable by default. Even when the user makes it executable the script fails due to a cloned repo. The syntax also didn't work when copied directly so I've included those changes to the README.md file.

* The install script isn't executable by default. Made it executable.

* Correcting Readme

I wanted to be better about the commands ran.